+++ runtime/onert/core/src/backend/builtin/KernelGenerator.cc
@@ -28,10 +28,11 @@ namespace onert::backend::builtin
 
 KernelGenerator::KernelGenerator(const ir::Graph &graph, DynamicTensorManager *dyn_tensor_manager,
                                  const std::shared_ptr<TensorRegistry> &tensor_reg,
+                                 const std::shared_ptr<custom::IKernelBuilder> &kernel_builder,
                                  const std::shared_ptr<ExternalContext> &external_context)
   : basic::KernelGeneratorBase{graph}, _dyn_tensor_manager{dyn_tensor_manager},
     _tensor_reg{tensor_reg}, _tensor_registries{}, _executors{nullptr}, _model_index{},
-    _external_context{external_context}
+    _kernel_builder{kernel_builder}, _external_context{external_context}
 {
   // DO NOTHING
 }
@@ -57,6 +58,33 @@ std::unique_ptr<exec::FunctionSequence> KernelGenerator::generate(ir::OperationI
   ret->append(std::move(_return_fn));
 
   return ret;
+}
+
+void KernelGenerator::visit(const ir::operation::Custom &node)
+{
+  auto fill_op_info = [&](const ir::OperandIndexSequence &opSeq,
+                          std::vector<custom::TypeInfo> &types,
+                          std::vector<IPortableTensor *> &tensors) {
+    for (const auto &idx : opSeq)
+    {
+      const auto &operand = _graph.operands().at(idx);
+      types.emplace_back(custom::TypeInfo{operand.shape(), operand.typeInfo().type()});
+      auto in_tensor = _tensor_reg->getPortableTensor(idx);
+      tensors.emplace_back(in_tensor);
+    }
+  };
+
+  backend::custom::CustomKernelConfigParams params{};
+
+  fill_op_info(node.getInputs(), params.input_types, params.input_tensors);
+  fill_op_info(node.getOutputs(), params.output_types, params.output_tensors);
+
+  params.userdata = node.userdata().data;
+  params.userdata_size = node.userdata().size;
+
+  auto fn = _kernel_builder->buildKernel(node.id(), std::move(params));
+
+  _return_fn = std::move(fn);
 }
 
 void KernelGenerator::visit(const ir::operation::Call &node)
